feat: add partial replacement of _.pick
While object destructuring has removed most of the need for a `pick` function, object
destructuring becomes rather unwieldy when trying to pick a large number of properties.
In that case, using a `pick` function makes the code more readable. Since underscore's
typing just returns `any`, adding a type safe `pick`.

NOTE: This doesn't support passing a function which determines the properties to pick as
underscore does. However, there doesn't seem to be a good way to make that behavior type
safe, so not attempting to support it at this time.